+++ e2sim/e2sm_examples/kpm_e2sm/src/kpm/encode_kpm.hpp
@@ -50,7 +50,7 @@ void encode_kpm_report_rancontainer_du(E2SM_KPM_IndicationMessage_t *indMsg);
 
 void encode_kpm_report_style1(E2SM_KPM_IndicationMessage_t* indicationmessage);
 
-void kpm_report_indication_header_initialized(E2SM_KPM_IndicationHeader_t *ihead, uint8_t *plmnid_buf, uint8_t *sst_buf, uint8_t *sd_buf, long fqival, long qcival, uint8_t *nrcellid_buf, uint8_t *gnbid_buf, int gnbid_unused, uint8_t *cuupid_buf, uint8_t *duid_buf, uint8_t *cuupname_buf);
+void kpm_report_indication_header_initialized(E2SM_KPM_IndicationHeader_t *ihead, uint8_t *plmnid_buf, uint8_t *sst_buf, uint8_t *sd_buf, long fqival, long qcival, uint8_t *nrcellid_buf, uint8_t *gnbid_buf, int gnbid_unused, int16_t *cuupid_buf, int16_t *duid_buf, uint8_t *cuupname_buf);
 
 void ue_meas_kpm_report_indication_message_initialized(E2SM_KPM_IndicationMessage_t* indicationmessage,uint8_t *nrcellid_buf,uint8_t *crnti_buf,const uint8_t *serving_buf, const uint8_t *neighbor_buf);
